chore: restructured lib modules
